Remove unnecessary interface
