+++ backend/src/main/java/de/bund/digitalservice/ris/search/repository/objectstorage/PortalBucket.java
@@ -5,7 +5,7 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 
 @Component
-public class PortalBucket extends S3ObjectStorage {
+public class PortalBucket extends ObjectStorage {
 
   public PortalBucket(@Qualifier("portalS3Client") ObjectStorageClient s3Client) {
     super(s3Client, LogManager.getLogger(PortalBucket.class));
