+++ backend/src/main/java/de/bund/digitalservice/ris/search/repository/objectstorage/literature/LiteratureBucketDummy.java
@@ -19,10 +19,10 @@ import org.springframework.stereotype.Component;
  */
 @Component
 @Profile({"uat", "production", "prototype"})
-public class LiteratureBucketDummy implements LiteratureBucket {
+public class LiteratureBucketDummy extends LiteratureBucket {
 
   public LiteratureBucketDummy() {
-    /**/
+    super(null, null);
   }
 
   public List<String> getAllKeys() {
